refactor(Packages): Moving MainActivity and ApplicationContext to activities and utils packages respectively
