+++ app/src/main/java/com/example/recipereviews/models/entities/User.java
@@ -7,7 +7,7 @@ import androidx.room.Entity;
 import androidx.room.Ignore;
 import androidx.room.PrimaryKey;
 
-import com.example.recipereviews.ApplicationContext;
+import com.example.recipereviews.utils.ApplicationContext;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.FieldValue;
