+++ app/src/main/java/com/example/recipereviews/utils/ImageUtil.java
@@ -13,7 +13,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageView;
 
-import com.example.recipereviews.ApplicationContext;
 import com.google.android.gms.common.util.Strings;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
